Fix PTH generation for standalone virtualenvs
Virtualenvs created in the runfiles tree can safely assume where they live- as such, they can happily use relative paths to find the correct location of whatever py_library targets they need to link in. However, using  is expected to create a virtualenv in the root source directory, presumably for user IDE integration. The existing code creates a broken one however- basically every path is changed to ../../../.. . Since this virtualenv doesn't live in a runfiles tree, it cannot find packages that way- but since it is running in 'bazel run' instead of during execution, we can use runfiles to find the correct absolute paths. I could not figure out a use case for strip_pth_workspace_root, or any uses of it that didn't result in a broken virtualenv, so I removed it.