feat: expose py_venv rule, allowing creation of virtual envs from multiple targets (#181)
Exposes the `py_venv` rule, which can create a Python virtual
environment from a set of targets.
This allows for easier creation of environments for IDE use, as they can
contain dependencies from multiple targets, such as tests.

The virtual environment that is created uses paths for first party
libraries that resolve into the source tree. While this allows for a
nicer user experience within the IDE, such as editing files being
"live", it is less "correct" as the Bazel target may reflect a different
state.