Remove BASH_BIN in favor of referencing bash via /usr/bin/env
While using rules_py in combination with rules_oci, I discovered an unexpected
behaviour: rules_py generates #! shebang lines using the path to bin that the
sh toolchain uses. This has the unexpected side effect of baking in the bash
path from the machine that built the container, which may or may not be the
same bash path that the container itself will have.

I discovered that macOS, Ubuntu, and the default @python docker image all have
bash at `/bin/bash`, but surprisingly, GitHub Actions workers have it at
`/usr/bin/bash`.

We can side step this problem entirely by using `#!/usr/bin/env bash` as our
shebang line, which I believe to be sufficiently portable and robust to use
for rules_py's scripts. See [this Stack Overflow](https://stackoverflow.com/questions/21612980/why-is-usr-bin-env-bash-superior-to-bin-bash)
question for further discussion on the approach.