Remove openssl (#269)
### Type of change

- Refactor (a code change that neither fixes a bug or adds a new
feature)


### Test plan

- Covered by existing test cases

#233 was mentioned as a fix for #218 but #233 includes a dependency on
openssl. This dependency on openssl makes building containers on MacOS
difficult as we then need to cross build openssl. I see that there was
an attempt to remove the openssl dependency `rattler_installs_packages`
however, `native-tls` is in the default features and is the only default
feature. Disabling `default-features` removes the dependency on openssl.

---------

Co-authored-by: Alex Eagle <alex@aspect.dev>