fix: use correct runfiles path for venv import paths
Fix for #276.

In `_make_import_path`, `ctx.workspace_name` is always `_main` when
evaluating the deps of targets in the main repo. This includes deps
that are in external repos, e.g.

    importer = use_extension("//:import.bzl", "importer")
    use_repo(importer, "myrepo")

with

    py_test(..., deps = ["@myrepo//foo"])

In this case, we actually want to use the label's `workspace_name`
attr which will be the same as the runfiles path that we need for the
venv import paths.

Also added a new e2e test that creates this external repo and verifies
the fix. Without the fix in place, the test does fail as described in
the e2e/repository-rule-deps/README.md "Repro" section.