chore: improve error message for py_runtime missing version info
When a python toolchain is added directly using `py_runtime`, etc, the
interpreter_version_info attribute (a dict) is required.

Prior to this change, it would fail in e.g. _py_binary_rule_impl with
an error:

    Error: 'dict' value has no field or method 'major'

Check this in earlier in _resolve_toolchain where we find the
"interpreter_version_info" attribute, ensuring that it has all of the
required dict keys.