feat: add coverage output to py_pytest_main
If the user runs with coverage enabled, Bazel sets an environment variable pointing to the files to be instrumented.
When that's present, attempt to import the coverage.py package, and use it to instrument and gather coverage, writing lcov directly.
Note the user must provide the coverage.py package and it must be version 6.3 or greater, so that lcov reporting is built-in.

Fixes #353