fix `java.io.IOException: Cannot run program ... Argument list too long` during pex generation.
In certain circumstance, a large number of `--source` args are passed to
the `_pex` script which causes an IOException to be thrown by bazel.
This refactors the rule to write the args to a file and pass that single
path as an arg to the `_pex` script. The script then opens that file and
feeds it into the arg parser instead.

fixes #542