fix(release): Configure the interpreter shim via a prebuilt toolchain (#568)
This fixes a breaking regression in v1.5.0 caused by the new interpreter
shim being referenced by source label and not via a toolchain which can
be supplied with prebuild artifacts.

Included in this change is a refactor consolidating our source of truth
for the list of prebuilt tools to configure down to `tools.bzl` which is
now shared with `tools/release` among other things, preventing how I got
myself into this mess.

As a further bit of paranoia this changeset also reworks the integration
tests so that instead of doing e2e testing with assets from a checkpoint
release (previously 1.1.0) we instead do a full build of the release
candidate assets, update the asset integrity file, run a local
fileserver hosting them and then do builds using the locally hosted
assets.

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes/no
- Breaking change (forces users to change their own code or config):
yes/no
- Suggested release notes appear below: yes/no

### Test plan

- [x] Existing test cases still green
- [x] Preexisting and inadequate e2e tests updated to cover this case