Fix debug logging in release (#613)
debug_build was enabled for users who were using
`--compilation_mode=dbg`, which lead to `set -x` being set in the
terminal used to create the venv, and tons of logging users don't need
to see. The help for the venv tool was also printed.

Replaces #596 which has conflicts.

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes/no
- Breaking change (forces users to change their own code or config):
yes/no
- Suggested release notes appear below: yes/no

### Test plan

- Covered by existing test cases
- New test cases added
- Manual testing; please provide instructions so we can reproduce:

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>