Support building PEX archives to inherit from the sys.path
By default PEX archives are built to operate fully hermetically, which
is usually a good idea. However, sometimes it is necessary to use
packages from the hosting environment. One such example is AWS Lambda;
the recommended approach for using PEX to run on Lambda is to build in
this way.[^1]

Mapping of this flag through from the build system largely follows the
equivalent feature in Pants.[^2]

[^1]: https://github.com/pex-tool/lambdex/blob/main/MIGRATING.md
[^2]: https://www.pantsbuild.org/stable/reference/targets/pex_binary#inherit_path