feat(py_venv): activate-less interpreter
This patch reworks how the interpreter shim operates to eliminate the
data dependency on the `activate` script having been evaluated and move
the parameters which previously came via the shell environment.

Because the behavior of `activate` isn't actually specified we
previously assumed that it was safe to just extend `activate`. This
isn't true at all it turns out, as under normal operation the
`./bin/python` link and the behavior of the interpreter initialization
as implemented in `Modules/getpath.py` provide special behavior when the
interpreter's observed filename (`argv[0]`) is a symlink as it is in a
normal non-relocatable virtualenv and will search for a `pyvenv.cfg` and
implicitly activate the venv as needed in that case.

Our `./bin/python` being an actual binary tool interrupts this which
means we're on our own to make `getpath` and `site` do the needful.

This should be as simple as setting the `home=` key in the `pyvenv.cfg`,
but there isn't a principled way other than going through the runfiles
library to locate where the interpreter lands in the runfiles tree
relative to the virtualenv. Ideally we'd set the `home=` key to a
relative path and mostly move on. Instead due in part to bzlmod munging
we have to do the repo mapping dance which can't be done statically. So
we need to do `rlocation` somewhere.

The solution this patch comes to is moving that `rlocation` logic out of
the `activate` script and into the launcher itself so that the launcher
can search for the needed `$PYTHONHOME` and desired interpreter without
the user having previously loaded the required parameters into the env
via `activate`.

Fixes #594 at the cost of making the shim binary unusable outside the
context of `rules_py`.