chore(ci): Automate pre-commit runs (#585)
This is an opinionated change, and to be very clear the opinion is mine.
Linters which can also do formatting _should not_ be allowed to break
CI. If there is a programmatically defined "right way" things should be
done and tooling can do it, no purpose is served in making a human rerun
a tool to apply that formatting.

Extend the `pre-commit` CI step so that it will commit (as
@aspect-marvin) any formatting changes and push them back to the branch.

The motivation for doing this isn't just my own annoyance with pushing
lint commits, but also so that PRs from @renovate-bot will trigger
updates to the Gazelle manifest (or other code) so that those PRs can
auto-merge.

---

### Changes are visible to end-users: no

### Test plan

N/A.

---------

Co-authored-by: aspect-marvin[bot] <marvin@aspect.build>