fix(py_venv): Partially resolve PYTHONHOME
Use the partially resolved interpreter path (eg. the first partially
resolved path for the interpreter binary which is under `.runfiles/`) as
the basis from which to identify the `pyvenv.cfg` file and thus as the
`sys.prefix`.

This fixes an issue which would allow the interpreter to observe an
un-resolved `sys.prefix` relative to which `_aspect.pth` source imports
would be invalid.