Remove one of the dependencies on uv-python (#695)
At first I thought this was the only dep and was excited to drop it, but
then I realized we also use it to detect the interpreter. Still, it
seems reasonable to avoid parsing an already-parsed version string and
this leaves us 1 step closers to trimming the dep.

While I was here, I prepped the action for path mapping so we could
avoid duplicating it across configs

### Changes are visible to end-users: no

### Test plan
existing tests?