fix(activate): Don't set PYTHONEXECUTABLE
Paired with @alexeagle. It used to be before recent improvements to
runfiles tree identification and the venv shim that we relied on the
activate script to set PYTHONEXECUTABLE. This proved to be unreliable,
in part because many tools just don't evaluate the activate script.

Setting the PYTHONEXECUTABLE without also setting PYTHONHOME breaks
non-Bazel Python invocations and so is a bug. Furthermore we didn't
save/restore or more importantly unset PYTHONEXECUTABLE during
deactivate, which broke shell environments.

Just don't set this.

Fixes #671.