fix(activate): Don't set PYTHONEXECUTABLE (#699)
Paired with @alexeagle. It used to be before recent improvements to
runfiles tree identification and the venv shim that we relied on the
activate script to set PYTHONEXECUTABLE. This proved to be unreliable,
in part because many tools just don't evaluate the activate script.

Setting the PYTHONEXECUTABLE without also setting PYTHONHOME breaks
non-Bazel Python invocations and so is a bug. Furthermore we didn't
save/restore or more importantly unset PYTHONEXECUTABLE during
deactivate, which broke shell environments.

Just don't set this.

Fixes #671.

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: yes

Fixed #671, `activate` scripts no longer poison the shell env.

### Test plan

- Manual testing

---------

Co-authored-by: aspect-marvin[bot] <marvin@aspect.build>