Fix absolute label use in main attribute of py_test (#582)
Replaces #463 from @keith. We should get this in as part of our next
break.

In rules_python you can specify a file from another package like:

```bzl
main = "//path/to:file.py",
```

This didn't work in rules_py since it was taking the main as a string to
manipulate. This appears closer to me to the upstream logic that was
mirrored, so maybe this was changed for another reason

---

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes/no
- Breaking change (forces users to change their own code or config):
yes/no
- Suggested release notes appear below: yes/no

### Test plan

- [x] New test cases added

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>